requirements.txt
google-genai
aiofiles
pillow
numpy
tqdm
json_repair
anthropic
openai
matplotlib
pyyaml
huggingface_hub
httpx
fastapi
uvicorn
jinja2
itsdangerous
python-multipart
sqlalchemy
alembic
pymysql
cryptography
